% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltCurves-input.R
\name{meltCurvesInput}
\alias{meltCurvesInput}
\title{Create a PCR curves viewer input control}
\usage{
meltCurvesInput(inputId, label = NULL, meltCurves, ggplotCode = NULL,
  selected = NULL, colorBy = "none", shapeBy = "none",
  diffCurves = TRUE, inverted = FALSE, showTm = TRUE,
  cssFile = system.file("/css/meltCurvesInputStyle.css", package =
  "shinyMolBio"), cssText = NULL, interactive = base::interactive())
}
\description{
Create an input control for representing PCR plate and dynamically selecting
wells inside it.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
ui <- fluidPage(
    pcrPlateInput("plate1",
                   "Plate 1",
                   RDML$new(system.file("/extdata/stepone_std.rdml", package = "RDML"))$AsTable(),
                  pcrFormatType$new(8,12,labelFormatType$new("ABC"),
                                         labelFormatType$new("123"))),
   verbatimTextOutput("selected")
 )
 server <- function(input, output) {
   output$selected <- renderText({ input$plate1 })
 }
 shinyApp(ui, server)
}
}
\seealso{
\code{\link{updatePcrPlateInput}}

Other input elements: \code{\link{pcrPlateInput}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\concept{input elements}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
