% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrCurves-input.R
\name{pcrCurvesInput}
\alias{pcrCurvesInput}
\title{Create a PCR curves viewer input control}
\usage{
pcrCurvesInput(inputId, label = NULL, pcrCurves, selected = NULL,
  ggplotCode = NULL, colorBy = NULL, shapeBy = NULL, logScale = FALSE,
  showCq = FALSE, showBaseline = FALSE,
  cssFile = system.file("/css/pcrCurvesInputStyle.css", package =
  "shinyMolBio"), cssText = NULL, interactive = base::interactive())
}
\description{
Create an input control for viewing PCR curves as plot.
}
\seealso{
\code{\link{updatePcrCurvesInput}}

Other input elements: \code{\link{meltCurvesInput}},
  \code{\link{pcrPlateInput}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
