% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderADplot.R
\name{print.adPlot}
\alias{print.adPlot}
\title{Printing adPlot}
\usage{
## S3 method for class 'adPlot'
print(x)
}
\arguments{
\item{x}{object of class \code{adPlot}}
}
\description{
Print a \code{adPlot}
}
\seealso{
\code{\link{renderAmpCurves}}, \code{\link{renderMeltCurves}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
