% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderCurves.R
\name{print.pcrCurves}
\alias{print.pcrCurves}
\title{Printing pcrCurves}
\usage{
## S3 method for class 'pcrCurves'
print(x)
}
\arguments{
\item{x}{object of class \code{pcrCurves}}
}
\description{
Print a \code{pcrCurves}
}
\examples{
library(RDML)
rdml <- RDML$new(system.file("/extdata/test.rdml", package = "shinyMolBio"))
curves <- renderMeltCurves("curves1", meltCurves = rdml$GetFData(dp.type = "mdp",
 long.table = TRUE))
curves
}
\seealso{
\code{\link{renderAmpCurves}}, \code{\link{renderMeltCurves}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
