% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrPlate-input.R
\name{print.pcrPlate}
\alias{print.pcrPlate}
\title{Printing PCR plate input control}
\usage{
## S3 method for class 'pcrPlate'
print(x)
}
\arguments{
\item{x}{object of class \code{pcrPlate}}
}
\description{
Print a \code{pcrPlate}
}
\seealso{
\code{\link{pcrPlateInput}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
