% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderADplot.R
\name{renderADplot}
\alias{renderADplot}
\title{Renders an allelic discrimination viewer}
\usage{
renderADplot(inputId, label = NULL, adData,
        targetColumn = "target.dyeId", xAxisTarget = "FAM",
        yAxisTarget = "HEX", valueColumn = "endPointRFU",
        colorBy = "genotype", polar = FALSE, showLegend = FALSE,
        plotlyCode = NULL, cssFile = NULL, cssText = NULL,
        interactive = TRUE)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to modify plot.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{adData}{Allelic descrimination data with
\code{RDML$AsTable()} format.}

\item{targetColumn}{Column name that contains axis splitting (dye or target).}

\item{xAxisTarget}{X axis target (dye) name.}

\item{yAxisTarget}{X axis target (dye) name.}

\item{valueColumn}{Column name that contains discrimination value (RFU or cq).}

\item{colorBy}{Column name that contains color levels data.}

\item{polar}{Enables polar coordinates.}

\item{showLegend}{Show plot legend.}

\item{plotlyCode}{Your quoted custom plotly code.}

\item{cssFile}{Path to the css styles file.}

\item{cssText}{CSS styles as text.}

\item{interactive}{Should be this \code{adPLot} interactive or not.}
}
\description{
Renders a reactive allelic discrimination plot that is suitable for
assigning to an \code{UI output} slot.
}
\examples{
library(RDML)
rdml <- RDML$new(system.file("/extdata/test.rdml", package = "shinyMolBio"))
tbl <- rdml$AsTable(endPointRFU = mean(tail(data$adp$fpoints$fluor, 5)))
tbl <- group_by(tbl, position)
tbl <- mutate(tbl,
              genotype = paste(
                if (endPointRFU[1] > 400 && endPointRFU[2] > 400) "AG"
                else if (endPointRFU[1] > 400) "AA"
                else if (endPointRFU[2] > 400) "GG"
                else "NA"
              ))
renderADplot("f", "aa", tbl, polar = TRUE, showLegend = TRUE)
}
\seealso{
\code{\link{updateADplot}}

Other render elements: 
\code{\link{renderAmpCurves}()},
\code{\link{renderMeltCurves}()}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\concept{render elements}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
