% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderCurves.R
\name{renderAmpCurves}
\alias{renderAmpCurves}
\title{Renders a PCR curves viewer input control}
\usage{
renderAmpCurves(inputId, label = NULL, pcrCurves, colorBy = NULL,
  shapeBy = NULL, logScale = FALSE, showCq = FALSE,
  showBaseline = FALSE,
  cssFile = system.file("/css/pcrCurvesInputStyle.css", package =
  "shinyMolBio"), cssText = NULL, interactive = base::interactive())
}
\description{
Create an input control for viewing PCR curves as plot.
}
\examples{
# TODO
}
\seealso{
\code{\link{updatePcrCurvesInput}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\concept{render elements}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
