% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderCurves.R
\name{renderMeltCurves}
\alias{renderMeltCurves}
\title{Renders a melting curves viewer}
\usage{
renderMeltCurves(inputId, label = NULL, meltCurves, colorBy = NULL,
  shapeBy = NULL, showTm = FALSE,
  cssFile = system.file("/css/curvesInputStyle.css", package =
  "shinyMolBio"), cssText = NULL, interactive = base::interactive())
}
\description{
Create melting curves plot.
}
\examples{
# TODO
}
\seealso{
\code{\link{updateCurvesInput}}

Other render elements: \code{\link{renderAmpCurves}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\concept{render elements}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
