% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shinyMolBio Example Applications}
\usage{
runExample(example)
}
\arguments{
\item{example}{The name of the example to run, or NA (the default) to list the available examples.}
}
\description{
Launch shinyMolBio example applications
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
 # List all available examples
 runExample()

 # Run one of the examples
 runExample("pcrPlateInput")

 # Print the directory containing the code for all examples
 system.file("shiny-examples", package="shinyMolBio")
}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
