% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderCurves.R
\name{updateCurves}
\alias{updateCurves}
\title{Change the value of a render PCR curves control on the client}
\usage{
updateCurves(session, inputId, label = NULL, hideCurves = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the \code{input} object.}

\item{label}{The label to set for the input object.}

\item{hideCurves}{The \code{fdata.names} of the curves to be hiden.}
}
\description{
Change the value of a render PCR curves control on the client
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
