% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrCurves-input.R
\name{updatePcrCurvesInput}
\alias{updatePcrCurvesInput}
\title{Change the value of a PCR curve input control on the client}
\usage{
updatePcrCurvesInput(session, inputId, label = NULL, selection = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the \code{input} object.}

\item{label}{The label to set for the input object.}

\item{selection}{The positions of the wells to be selected.}
}
\description{
Change the value of a PCR plate input control on the client
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
