% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrCurves-input.R
\name{pcrCurvesInput}
\alias{pcrCurvesInput}
\title{Create a PCR curves viewer input control}
\usage{
pcrCurvesInput(inputId, label = NULL, pcrCurves, colorBy = "none",
  shapeBy = "none", logScale = FALSE, showCq = FALSE,
  showBaseline = FALSE,
  cssFile = system.file("/css/pcrCurvesInputStyle.css", package =
  "shinyMolBio"), cssText = NULL, plateLegend = NULL,
  interactive = base::interactive())
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the selected
wells positions.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{cssFile}{path to the css styles file.}

\item{cssText}{css styles as text.}

\item{plateLegend}{plate legend (any HTML content).}

\item{plateDescription}{Plate description - basicly output from \code{RDML
AsTable()} function.}

\item{pcrFormat}{PCR plate parametrs. Should be \code{pcrFormatType}.}

\item{selection}{Set preselected wells (e.g. \code{c("A01", "A02") or \code{c(1, 2)}})}

\item{wellLabelTemplate}{Template of the well label.}

\item{onHoverWellTextTemplate}{Template of the text on hover.}

\item{wellClassTemplate}{Template of the well class (css class).}

\item{wellStyleTemplate}{Template of the well style (css).}

\item{wellGroupTemplate}{Template of the well group for selecting.}
}
\value{
A PCR plate control that can be added to a UI definition.
}
\description{
Create an input control for representing PCR plate and dynamically selecting
wells inside it.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
ui <- fluidPage(
    pcrPlateInput("plate1",
                   "Plate 1",
                   RDML$new(system.file("/extdata/stepone_std.rdml", package = "RDML"))$AsTable(),
                  pcrFormatType$new(8,12,labelFormatType$new("ABC"),
                                         labelFormatType$new("123"))),
   verbatimTextOutput("selected")
 )
 server <- function(input, output) {
   output$selected <- renderText({ input$plate1 })
 }
 shinyApp(ui, server)
}
}
\seealso{
\code{\link{updatePcrPlateInput}}

Other input elements: \code{\link{pcrPlateInput}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
