% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrCurves-input.R
\name{pcrCurvesInput}
\alias{pcrCurvesInput}
\title{Create a PCR curves viewer input control}
\usage{
pcrCurvesInput(inputId, label = NULL, pcrCurves, selected = NULL,
  colorBy = "none", shapeBy = "none", logScale = FALSE, showCq = FALSE,
  showBaseline = FALSE,
  cssFile = system.file("/css/pcrCurvesInputStyle.css", package =
  "shinyMolBio"), cssText = NULL, interactive = base::interactive())
}
\description{
Create an input control for representing PCR plate and dynamically selecting
wells inside it.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
ui <- fluidPage(
    pcrPlateInput("plate1",
                   "Plate 1",
                   RDML$new(system.file("/extdata/stepone_std.rdml", package = "RDML"))$AsTable(),
                  pcrFormatType$new(8,12,labelFormatType$new("ABC"),
                                         labelFormatType$new("123"))),
   verbatimTextOutput("selected")
 )
 server <- function(input, output) {
   output$selected <- renderText({ input$plate1 })
 }
 shinyApp(ui, server)
}
}
\seealso{
\code{\link{updatePcrPlateInput}}

Other input elements: \code{\link{pcrPlateInput}}
}
\author{
Konstantin A. Blagodatskikh <k.blag@yandex.ru>
}
\keyword{Input}
\keyword{PCR}
\keyword{RDML}
\keyword{Shiny}
